+++ src/pages/Documents/index.jsx
@@ -1,10 +1,8 @@
-import React, { useEffect } from "react";
+import React from "react";
 import ListDocuments from "./components/ListDocuments";
 import UploadDocument from "./components/UploadDocument";
 
 const Index = () => {
-  useEffect(() => {}, []);
-
   return (
     <div className="px-40 pt-8">
       <UploadDocument />