+++ src/redux/alert/actions.js
@@ -1,0 +1,8 @@
+import { SET_ALERT } from "../actions";
+
+export const setAlert = (alert) => {
+  return {
+    type: SET_ALERT,
+    payload: { alert },
+  };
+};