+++ src/tests/pages/components/Logo.test.js
@@ -4,6 +4,7 @@ import Logo from "../../../pages/components/Logo";
 describe("<Logo />", () => {
   test("match snapshot", () => {
     const wrapper = shallow(<Logo />);
+    // render logo corrently
     expect(wrapper).toMatchSnapshot();
   });
 });